% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{activator_significant}
\alias{activator_significant}
\title{activator_significant}
\usage{
activator_significant(
  hub.int.significant = hub.int.significant,
  network.int = network.int
)
}
\arguments{
\item{hub.int.significant}{significance of the hub genes}

\item{network.int}{netowrk edges}
}
\value{
res
}
\description{
define the activator significance
}
