% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_wrappers.R
\name{compute_TSI_TSE_parallel}
\alias{compute_TSI_TSE_parallel}
\title{compute_TSI_TSE_parallel (C++)}
\usage{
compute_TSI_TSE_parallel(expression)
}
\arguments{
\item{expression}{Numeric expression matrix.}
}
\value{
A list containing TSI and TSE values.
}
\description{
Internal function called by NECorr to compute Tissue Specificity Index (TSI) and Tissue Specific Expression (TSE) in parallel.
}
\keyword{internal}
