% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationGini.R
\name{cor.matrix.NECorr}
\alias{cor.matrix.NECorr}
\title{cor.matrix.NECorr}
\usage{
cor.matrix.NECorr(
  GEMatrix,
  cpus = cpus,
  cormethod = c("GCC", "PCC", "SCC", "KCC"),
  style = c("pairs.between", "pairs.only"),
  var1.id = NA,
  var2.id = NA,
  pernum = 0,
  sigmethod = c("two.sided", "one.sided"),
  output = c("matrix", "paired")
)
}
\arguments{
\item{GEMatrix}{expression matrix}

\item{cpus}{number of cpu used}

\item{cormethod}{correlation method:"GCC", "PCC", "SCC", "KCC"}

\item{style}{type of pairs: "pairs.between","pairs.only"}

\item{var1.id}{variable vector source}

\item{var2.id}{varaible vector target}

\item{pernum}{permutation number}

\item{sigmethod}{significance}

\item{output}{matrix or paired}
}
\value{
resList
}
\description{
cor.matrix.NECorr
}
