% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.ranking.R
\name{DE.ranking}
\alias{DE.ranking}
\title{DE.ranking: Differential expression ranking for NECorr}
\usage{
DE.ranking(exps, GeneList, factortab, sample.l, sample.names)
}
\arguments{
\item{exps}{Expression matrix (genes × samples)}

\item{GeneList}{Vector of gene IDs to include}

\item{factortab}{Factor vector or data frame of sample groups}

\item{sample.l}{Condition of interest}

\item{sample.names}{Vector of all sample group names}
}
\value{
Named vector of scaled DE ranks
}
\description{
Uses limma to rank genes in the network by differential expression.
}
