% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueSpecificity.R
\name{DiscNULLrow}
\alias{DiscNULLrow}
\title{DiscNULLrow}
\usage{
DiscNULLrow(eset)
}
\arguments{
\item{eset}{entry table}
}
\value{
res table with null column
}
\description{
Discard rows of expression with only NULL expression
}
