% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{effector_edge_significance}
\alias{effector_edge_significance}
\title{effector_edge_significance}
\usage{
effector_edge_significance(
  network.int = network.int,
  gene.rank.eff.hash = gene.rank.eff.hash
)
}
\arguments{
\item{network.int}{network data}

\item{gene.rank.eff.hash}{gene ranking for the effector in a hash}
}
\value{
res
}
\description{
define the edge ranking around the effectors
}
