% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{effector_significance}
\alias{effector_significance}
\title{effector_significance}
\usage{
effector_significance(eff.m.param, Desc, j.nB)
}
\arguments{
\item{eff.m.param}{table with scale parameters for gene ranking}

\item{Desc}{description table of the genes}

\item{j.nB}{predictive model based on Naives Bayes}
}
\value{
res list of 1.hash effector gene ranking and 2.table of ranking and descriptions
}
\description{
define the ranking of the effector genes
}
