% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_weight.R
\name{entropy_weight}
\alias{entropy_weight}
\title{entropy_weight}
\usage{
entropy_weight(df)
}
\arguments{
\item{df}{A data table or matrix with numeric values.}
}
\value{
A numeric vector of entropy weights for each column.
}
\description{
Calculate the entropy weight for each column in a data frame or matrix.
}
