% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationGini.R
\name{expand.grid.unique}
\alias{expand.grid.unique}
\title{expand.grid.unique}
\usage{
expand.grid.unique(x, y, include.equals = FALSE)
}
\arguments{
\item{x}{x parameter}

\item{y}{y parameter}

\item{include.equals}{including equals}
}
\value{
res
}
\description{
expand the data
}
