% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iut_functions.R
\name{fast_tsi_tse}
\alias{fast_tsi_tse}
\title{fast_tsi_tse}
\usage{
fast_tsi_tse(expr, tissues)
}
\arguments{
\item{expr}{Numeric expression matrix (genes x samples).}

\item{tissues}{Character vector of tissue types for each sample (column).}
}
\value{
Data frame with columns: gene, TSI, TSE (normalized from 0 to 1).
}
\description{
Fast computation of Tissue Specificity Index (TSI) and Tissue Specific Expression (TSE)
for genes across tissues, handling replicates and normalizing results. Collapse replicates,
run parallel TSI/TSE, and normalise from 0 to 1
}
