% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationGini.R
\name{gcc.corfinal}
\alias{gcc.corfinal}
\title{gcc.corfinal}
\usage{
gcc.corfinal(gcccor)
}
\arguments{
\item{gcccor}{table of the rank gcc}
}
\value{
res table of gcc score and gcc pvalue
}
\description{
gcc.corfinal
}
