% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationGini.R
\name{gcc}
\alias{gcc}
\title{gcc}
\usage{
gcc(weightvec, vectsort, vectselfsort)
}
\arguments{
\item{weightvec}{weight}

\item{vectsort}{vector sorted}

\item{vectselfsort}{selection of sorted}
}
\value{
gcccor
}
\description{
fom package XXX - reference
}
