% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationGini.R
\name{getcor}
\alias{getcor}
\title{getcor}
\usage{
getcor(g1, g2, cormethod)
}
\arguments{
\item{g1}{gcc first}

\item{g2}{gcc second}

\item{cormethod}{method}
}
\value{
res
}
\description{
from package XXX - reference
}
