% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationGini.R
\name{getpvalue}
\alias{getpvalue}
\title{getpvalue}
\usage{
getpvalue(percorvec, pernum, realcor, sigmethod)
}
\arguments{
\item{percorvec}{percentage corelation vector}

\item{pernum}{permutaion number}

\item{realcor}{real correlation}

\item{sigmethod}{significance method}
}
\value{
pvalue
}
\description{
from package XXX - reference
}
