% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gini}
\alias{gini}
\title{Do values in a numeric vector fall in specified range?}
\usage{
gini(edges, expression, bootstrapIterations, statCutoff)
}
\arguments{
\item{edges}{dataframe with the edges with 3 columns: source, interaction and
target}

\item{expression, }{expression table containing the genes}

\item{bootstrapIterations}{number of bootstrap}

\item{statCutoff}{stat cut-off p-value}
}
\description{
This is a shortcut for `x >= left & x <= right`, implemented
efficiently in C++ for local values, and translated to the
appropriate SQL for remote tables.
}
