% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationGini.R
\name{giniR}
\alias{giniR}
\title{GiniR}
\usage{
giniR(edges, expression, bootstrapIterations, statCutoff)
}
\arguments{
\item{edges}{the network edges}

\item{expression}{the expression tables}

\item{bootstrapIterations}{the parameter of boothtrap for randomization}

\item{statCutoff}{significance cut off}
}
\description{
convertion of Rcpp into R function
}
