% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{hub_edge_significance}
\alias{hub_edge_significance}
\title{hub_edge_significance}
\usage{
hub_edge_significance(
  network.int = network.int,
  gene.rank.hash = gene.rank.hash
)
}
\arguments{
\item{network.int}{network file}

\item{gene.rank.hash}{gene ranking hash}
}
\value{
res
}
\description{
define the hub edge ranking
}
