% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indexing.network}
\alias{indexing.network}
\title{indexing.network}
\usage{
indexing.network(tab, network)
}
\arguments{
\item{tab}{Expression matrix (rownames = genes)}

\item{network}{Two-column data frame of edges (source, target)}
}
\value{
Matrix of same size as network with integer indices
}
\description{
Replace gene names in network with their row indices in expression matrix
}
