% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueSpecificity.R
\name{IUTest}
\alias{IUTest}
\title{IUTest}
\usage{
IUTest(m.eset, sFactors, sIndex, alpha)
}
\arguments{
\item{m.eset}{expression matrix}

\item{sFactors}{number of sample replicates}

\item{sIndex}{index of the sample in the table}

\item{alpha}{significance level for the IUT}
}
\value{
res
}
\description{
perform the complete IUT for a gene expression table
}
