% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iut_functions.R
\name{IUTtse}
\alias{IUTtse}
\title{IUTtse}
\usage{
IUTtse(
  data,
  geneTS,
  targetSample = NULL,
  maxExpThreshold = NULL,
  tissues = NULL
)
}
\arguments{
\item{data}{Expression matrix (genes x tissues).}

\item{geneTS}{Logical vector marking tissue-excluded genes.}

\item{targetSample}{Name of the target tissue/condition.}

\item{maxExpThreshold}{Maximum expression threshold for targetSample.}

\item{tissues}{Optional vector of tissue types for columns (for handling replicates).}
}
\value{
Data frame with \code{tse.order} as last column.
}
\description{
Define tissue-excluded genes from the tissue/condition of interest.
}
