% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linked_act_hub_net}
\alias{linked_act_hub_net}
\title{linked_act_hub_net}
\usage{
linked_act_hub_net(hub.int.significant, gene.rank.act.significant, network.int)
}
\arguments{
\item{hub.int.significant}{significant network of the hub genes}

\item{gene.rank.act.significant}{the ranking of the activator genes}

\item{network.int}{full gene network}
}
\value{
act.net
}
\description{
linked_act_hub_net
}
