% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{linked_eff_hub_net}
\alias{linked_eff_hub_net}
\title{linked_eff_hub_net}
\usage{
linked_eff_hub_net(
  hub.int.significant = hub.int.significant,
  eff.int.significant = eff.int.significant
)
}
\arguments{
\item{hub.int.significant}{significant network of the hub genes}

\item{eff.int.significant}{significant network of the effector genes}
}
\value{
eff.net
}
\description{
linked_eff_hub_net
}
