% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationGini.R
\name{multiCorr}
\alias{multiCorr}
\title{multiCorr}
\usage{
multiCorr(
  x,
  net = NA,
  nsockets = 4,
  methods = c("GCC", "PCC", "SCC", "KCC"),
  sigmethod = c("two.sided", "one.sided"),
  nblocks = 10,
  verbose = TRUE,
  cpus = 1,
  pernum = 0,
  ...
)
}
\arguments{
\item{x}{the table of expression}

\item{net}{the network to test for co-expression}

\item{nsockets}{number of parallel sockets}

\item{methods}{correlation method}

\item{sigmethod}{significance of the p-value for the chosen method}

\item{nblocks}{number of blocks to be cut}

\item{verbose}{enumerateall steps}

\item{cpus}{number of cpu used}

\item{pernum}{number of permutation for singificance test}

\item{...}{other parameters}
}
\value{
corMAT a correlation matrix
}
\description{
the function is calculating co-expression
}
