% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/necorr_local.R
\name{NECorr}
\alias{NECorr}
\alias{Necorr}
\title{Output standard human genome seqlengths}
\usage{
Necorr(network.file, expression, description.file, condition, metadata,
  name, Filelist, method = "GCC", permutation = 1000, sigcorr = 0.01,
  fadjacency = "only", type = "gene", dirout = "./", NSockets = 2)
}
\arguments{
\item{network.file}{Molecular network file with source in the first column, targets in
the second column}

\item{expression}{Expression file in log2 (ratio expression) with row: gene,
first column: type of sample,second column: sample names}

\item{description.file}{genome description}

\item{condition}{Condition from expression to study the network co-expression
correlation}

\item{method}{Method used for co-expression correlation: GCC, MINE, PCC,
SCC or KCC}

\item{permutation}{permutation number used for all significance calculation}

\item{fadjacency}{correlation with all combination (all) or network
combination only (only)}

\item{type}{Omics comparative expression type: protein or gene}

\item{dirout}{directory for the results}

\item{lmiR}{List of miRNAs}
}
\description{


NECorr helps discover candidate genes that could be
important for specific conditions.
The principal inputs are the expression data and the network file.
The expression data should start with 3 header columns.
The first column describes the conditions. Each condition will be
treated separately for the co-expression analysis
The output of the program will be generated in a result folder generated
in the working path
}
\author{
Christophe Liseron-Monfils, Andrew Olson
}
