% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\name{necorr_graph}
\alias{necorr_graph}
\title{necorr_graph}
\usage{
necorr_graph(hubnet, hub.int.significant, gene.rank.act.significant)
}
\arguments{
\item{hubnet}{significant network contining hub, activators}

\item{hub.int.significant}{hub interaction significance}

\item{gene.rank.act.significant}{activator ranking}
}
\value{
netgraph
}
\description{
necorr_graph
}
