% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetTopology.R
\name{NetTopology}
\alias{NetTopology}
\title{Network topology statistics}
\usage{
NetTopology(network)
}
\arguments{
\item{network}{Data frame with 'source' and 'target' columns}
}
\value{
Data frame of topology metrics for each gene
}
\description{
Calculates centrality and connectivity metrics for each gene in the network.
}
\details{
NetTopology
}
