% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaling_param}
\alias{scaling_param}
\title{scaling_param}
\usage{
scaling_param(x)
}
\arguments{
\item{x}{Named numeric vector}
}
\value{
Data frame with scaled values
}
\description{
Scale named numeric vector from 0 to 100
}
