% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueSpecificity.R
\name{ts.IUT}
\alias{ts.IUT}
\title{ts.IUT}
\usage{
ts.IUT(name = "exp", eset, tissues, target, threshold = 0.05, filter = 50)
}
\arguments{
\item{name}{name of the data}

\item{eset}{matrix of expression data in matrix format in TPM for example
but any other expression matrix}

\item{tissues}{character vector defining the the tissue type of each
column present in the expression table}

\item{target}{the target tissues}

\item{threshold}{represent the p-value for the IUT test after Sidak
correction; the outlier percentage for the ROKU test
the percentage of (1 -threshold for the Tindex and the TSI tests)}

\item{filter}{the minimal expression level for the tissue/condition of
interest; the default is set at 100; expression of 100 TPM}
}
\value{
list.all csv.table output for each type of analysis will be written in the
table folder
}
\description{
combine the TSI/TSE test and IUT
to determine the tissue specific genes and the tissue excluded genes.
list of list with 4 methods raw and filtered results matrix
of tissue specific genes for each method; upload all libraries gplots
TCC
upload function below
TissueTransform, DiscNULLrow, factOrg, IUTest, numeratorTSI, denominator,
TSI, tsi.matrix,
plotlOnegene, plotLfewGenes, heatmap.ts, tindex, tindex.matrix,
}
