% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueSpecificity.R
\name{TSE}
\alias{TSE}
\title{TSE}
\usage{
TSE(x)
}
\arguments{
\item{x}{vector of value needed to be "TSEed"}
}
\value{
res
}
\description{
tissue-selective downregulated genes (Liseron et al., 2013)
}
