% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{activator_significant}
\alias{activator_significant}
\title{activator_significant}
\usage{
activator_significant(hub.int.significant, network.int, Desc)
}
\arguments{
\item{hub.int.significant}{significance of the hub genes}

\item{network.int}{network edges}

\item{Desc}{description file genes and gene names}
}
\value{
res
}
\description{
define the activator significance
}
