% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{DE.ranking}
\alias{DE.ranking}
\title{DE.ranking}
\usage{
DE.ranking(
  exps,
  GeneList,
  factortab,
  sample.l,
  sample.names,
  exps.file = FALSE
)
}
\arguments{
\item{exps}{expression table}

\item{GeneList}{gene list present in the network}

\item{factortab}{table with factors}

\item{sample.l}{the sample chosen for tissue specificity}

\item{sample.names}{all the samples in the metadata}

\item{exps.file}{the expression file}
}
\value{
ranks rank of the gene in DEG analysis
}
\description{
differential expression ranking using the network gene
  1-expression file with NECorr format
  2-list of network genes
  3-expression factor table generate automatically
  4-name of the studied sample
  read expressiontable from microarrays or RNAseq
}
