% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueSpecificity.R
\name{factOrg}
\alias{factOrg}
\title{factOrg}
\usage{
factOrg(eset, pDataTissue)
}
\arguments{
\item{eset}{expression data}

\item{pDataTissue}{vector containing the sample names}
}
\value{
out from list the factor list of the sample type
}
\description{
reorganize the column from the list of Tissue
}
