% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fishersMethod}
\alias{fishersMethod}
\title{fishersMethod}
\usage{
fishersMethod(x)
}
\arguments{
\item{x}{Numeric vector of p-values}
}
\value{
Combined p-value
}
\description{
Combine p-values using Fisher's method.
}
