% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{indexing.network}
\alias{indexing.network}
\title{indexing.network}
\usage{
indexing.network(tab, network)
}
\arguments{
\item{tab}{expression table}

\item{network}{network needed to be indexed}
}
\value{
netindex
}
\description{
indexing.network
}
