% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueSpecificity.R
\name{IUTtsi}
\alias{IUTtsi}
\title{IUTtsi}
\usage{
IUTtsi(data, geneTS, targetSample, expThreshold)
}
\arguments{
\item{data}{expression data}

\item{geneTS}{genes with tissue specific proprieties}

\item{targetSample}{target sample name}

\item{expThreshold}{threshold of the expression for RNAseq around 10 for microarray
could be different according to the platform}
}
\value{
y.sort
}
\description{
define tissue specificexpresed genes expThreshold expression threshold to apply to detect tissue-specificity
This could avoid very lowly expressed genes that could be artifact
}
