% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NECorr.R
\name{NECorr}
\alias{NECorr}
\title{NECorr}
\usage{
NECorr(
  networkFile,
  expression,
  description.file,
  condition,
  meta_data,
  permutation = 1000,
  method = c("GCC"),
  visualize = TRUE,
  output_dir = NULL,
  interactive_net = FALSE,
  top_n = 20,
  save_results = TRUE,
  useBestGCC = FALSE,
  asymmetricGCC = FALSE,
  ...
)
}
\arguments{
\item{networkFile}{file or data.table of Molecular network file with source in the first column,
targets in the second column
with the same name than the network file and the same localization}

\item{expression}{file or data.table of Expression file in log2 (ratio expression) with row: gene,
first column: type of sample,second column: sample names}

\item{description.file}{file or data.table of the genome description}

\item{condition}{Condition from expression to study the network co-expression correlation}

\item{meta_data}{file or data.table with the meta_data}

\item{permutation}{permutation number used for all significance calculation}

\item{method}{correlation method: GCC, PCC, KCC, SCC, SPP; c("GCC", "PCC", "SCC", "KCC")}

\item{visualize}{TRUE/FALSE whether to generate visualizations of the results}

\item{output_dir}{Optional directory to save results and plots if}

\item{interactive_net}{TRUE/FALSE whether to generate an interactive network}

\item{top_n}{Number of top hub genes to highlight in plots}

\item{save_results}{TRUE/FALSE whether to save results to CSV files}

\item{useBestGCC}{TRUE/FALSE whether to use the best GCC method}

\item{asymmetricGCC}{TRUE/FALSE whether to use asymmetric GCC, only if useBestGCC is TRUE, see multiCorr documentation}

\item{...}{Additional arguments passed to correlation functions.}
}
\value{
res
}
\description{
NECorr helps discover candidate genes that could be
important for specific conditions.
The principal inputs are the expression data and the network file.
The expression data should start with 3 header columns.
The first column describes the conditions. Each condition will be
treated separately for the co-expression analysis
The output of the program will be generated in a result folder generated
in the working path
Create the output directory if not existing; generate "./results" dir and
"./results/tmp"
C.Liseron-Monfils - Ware lab Sept2013 - CSHL
partly based on rsgcc package for the GCC, PCC,KCC and SCC correlation
Ma et al, 2012, plant Physiology
}
\author{
Christophe Liseron-Monfils, Andrew Olson
}
