% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{NetTopology}
\alias{NetTopology}
\title{NetTopology}
\usage{
NetTopology(network)
}
\arguments{
\item{network}{the network should be a list of interactions space by a line}
}
\value{
datagrid
}
\description{
network topology statistics using igraph methods;
script partly adapted from https://gist.github.com/mhawksey/16-2306
}
\author{
Christophe Liseron-Monfils, Martin Hawksey
}
