% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ScalN}
\alias{ScalN}
\title{ScalN}
\usage{
ScalN(x, ...)
}
\arguments{
\item{x}{vector that need to be rescaled}

\item{...}{other paramter that can be passed on}
}
\value{
res
}
\description{
rescaling data in a range between 0 to 1
}
