% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iut_functions.R
\name{TSE}
\alias{TSE}
\title{TSE}
\usage{
TSE(x)
}
\arguments{
\item{x}{Numeric vector of expression values for a gene across tissues.}
}
\value{
Numeric TSE value
}
\description{
Tissue Specificity Exclusion Index (Liseron et al., 2013) for downregulated genes.
This measures how much a gene is excluded in one tissue compared to others.
}
