% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueSpecificity.R
\name{TSI}
\alias{TSI}
\title{TSI}
\usage{
TSI(x)
}
\arguments{
\item{x}{vector of value needed to be "TSIed"}
}
\value{
res vector scaled to the TSI
}
\description{
tissue-selective upregulated genes from Tissue Specificty index (Yanai et al, 2005)
}
